Move svp fund transaction tests to new FederationChangeTest class
