+++ rskj-core/src/test/java/co/rsk/peg/BridgeSupportTest.java
@@ -543,271 +543,6 @@ class BridgeSupportTest {
     }
 
     @Nested
-    @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-    @Tag("svp tests")
-    class SvpTests {
-        private final Coin spendableValueFromProposedFederation = bridgeMainNetConstants.getSpendableValueFromProposedFederation();
-        private final List<LogInfo> logs = new ArrayList<>();
-
-        private BridgeStorageProvider bridgeStorageProvider;
-
-        private BridgeSupport bridgeSupport;
-
-        private Federation activeFederation;
-        private Federation proposedFederation;
-
-        private BtcTransaction svpFundTransactionUnsigned;
-        private Sha256Hash svpFundTransactionHashUnsigned;
-
-        private Transaction rskTx;
-        private Keccak256 rskTxHash;
-        private Block rskExecutionBlock;
-
-        @BeforeEach
-        void setUp() {
-            long rskExecutionBlockNumber = 1000L;
-            long rskExecutionBlockTimestamp = 10L;
-            BlockHeader blockHeader = new BlockHeaderBuilder(mock(ActivationConfig.class))
-                .setNumber(rskExecutionBlockNumber)
-                .setTimestamp(rskExecutionBlockTimestamp)
-                .build();
-            rskExecutionBlock = Block.createBlockFromHeader(blockHeader, true);
-
-            activeFederation = FederationTestUtils.getErpFederation(federationConstantsMainnet.getBtcParams());
-            List<UTXO> activeFederationUTXOs = BitcoinTestUtils.createUTXOs(10, activeFederation.getAddress());
-            proposedFederation = P2shErpFederationBuilder.builder().build();
-
-            federationSupport = mock(FederationSupport.class);
-            when(federationSupport.getActiveFederation()).thenReturn(activeFederation);
-            when(federationSupport.getActiveFederationAddress()).thenReturn(activeFederation.getAddress());
-            when(federationSupport.getActiveFederationBtcUTXOs()).thenReturn(activeFederationUTXOs);
-            when(federationSupport.getProposedFederation()).thenReturn(Optional.of(proposedFederation));
-
-            Coin feePerKb = Coin.valueOf(1000);
-            when(feePerKbSupport.getFeePerKb()).thenReturn(feePerKb);
-
-            rskTxHash = PegTestUtils.createHash3(1);
-            rskTx = mock(Transaction.class);
-            when(rskTx.getHash()).thenReturn(rskTxHash);
-
-            BridgeEventLogger bridgeEventLogger = new BridgeEventLoggerImpl(
-                bridgeMainNetConstants,
-                activationsAfterForks,
-                logs,
-                signatureCache
-            );
-
-            Repository repository = createRepository();
-            bridgeStorageProvider = new BridgeStorageProvider(
-                repository,
-                contractAddress,
-                btcMainnetParams,
-                activationsAfterForks
-            );
-
-            bridgeSupport = bridgeSupportBuilder
-                .withBridgeConstants(bridgeMainNetConstants)
-                .withProvider(bridgeStorageProvider)
-                .withEventLogger(bridgeEventLogger)
-                .withActivations(activationsAfterForks)
-                .withFederationSupport(federationSupport)
-                .withFeePerKbSupport(feePerKbSupport)
-                .withExecutionBlock(rskExecutionBlock)
-                .build();
-        }
-
-        @Test
-        void processSvpFundTransactionWithoutSignatures_whenProposedFederationDoesNotExist_throwsIllegalStateException() {
-            // arrange
-            when(federationSupport.getProposedFederation()).thenReturn(Optional.empty());
-
-            // act & assert
-            IllegalStateException exception = assertThrows(IllegalStateException.class,
-                () -> bridgeSupport.processSvpFundTransactionUnsigned(rskTx));
-
-            String expectedMessage = "Proposed federation should be present when processing SVP fund transaction.";
-            assertEquals(expectedMessage, exception.getMessage());
-        }
-
-        @Test
-        void processSvpFundTransactionWithoutSignatures_whenThereAreNoEnoughUTXOs_throwsInsufficientMoneyException() {
-            // arrange
-            List<UTXO> insufficientUtxos = new ArrayList<>();
-            when(federationSupport.getActiveFederationBtcUTXOs()).thenReturn(insufficientUtxos);
-
-            // act & assert
-            InsufficientMoneyException exception = assertThrows(InsufficientMoneyException.class,
-                () -> bridgeSupport.processSvpFundTransactionUnsigned(rskTx));
-
-            String totalNeededValueInBtc = spendableValueFromProposedFederation.multiply(2).toFriendlyString();
-            String expectedMessage = String.format("Insufficient money,  missing %s", totalNeededValueInBtc);
-            assertEquals(expectedMessage, exception.getMessage());
-        }
-
-        @Test
-        void processSvpFundTransactionWithoutSignatures_createsExpectedTransactionAndSavesTheHashInStorageEntryAndPerformsPegoutActions() throws Exception {
-            // act
-            bridgeSupport.processSvpFundTransactionUnsigned(rskTx);
-            bridgeStorageProvider.save(); // to save the tx sig hash
-
-            // assert
-            assertSvpFundTxHashUnsignedWasSavedInStorage();
-            assertSvpReleaseWasSettled();
-            assertSvpFundTransactionHasExpectedInputsAndOutputs();
-        }
-
-        private void assertSvpFundTxHashUnsignedWasSavedInStorage() {
-            Optional<Sha256Hash> svpFundTransactionHashUnsignedOpt = bridgeStorageProvider.getSvpFundTxHashUnsigned();
-            assertTrue(svpFundTransactionHashUnsignedOpt.isPresent());
-
-            svpFundTransactionHashUnsigned = svpFundTransactionHashUnsignedOpt.get();
-        }
-
-        private void assertSvpReleaseWasSettled() throws IOException {
-            PegoutsWaitingForConfirmations pegoutsWaitingForConfirmations = bridgeStorageProvider.getPegoutsWaitingForConfirmations();
-            assertPegoutWasAddedToPegoutsWaitingForConfirmations(pegoutsWaitingForConfirmations, svpFundTransactionHashUnsigned, rskTxHash);
-
-            svpFundTransactionUnsigned = getSvpFundTransactionFromPegoutsMap(pegoutsWaitingForConfirmations);
-
-            assertPegoutTxSigHashWasSaved(svpFundTransactionUnsigned);
-            assertLogReleaseRequested(rskTxHash, svpFundTransactionHashUnsigned, spendableValueFromProposedFederation);
-            assertLogPegoutTransactionCreated(svpFundTransactionUnsigned);
-        }
-
-        private BtcTransaction getSvpFundTransactionFromPegoutsMap(PegoutsWaitingForConfirmations pegoutsWaitingForConfirmations) {
-            Set<PegoutsWaitingForConfirmations.Entry> pegoutEntries = pegoutsWaitingForConfirmations.getEntries();
-            assertEquals(1, pegoutEntries.size());
-            // we now know that the only present pegout is the fund tx
-            Iterator<PegoutsWaitingForConfirmations.Entry> iterator = pegoutEntries.iterator();
-            PegoutsWaitingForConfirmations.Entry pegoutEntry = iterator.next();
-
-            return pegoutEntry.getBtcTransaction();
-        }
-
-        private void assertSvpFundTransactionHasExpectedInputsAndOutputs() {
-            assertInputsAreFromActiveFederation();
-
-            List<TransactionOutput> svpFundTransactionUnsignedOutputs = svpFundTransactionUnsigned.getOutputs();
-            int svpFundTransactionUnsignedOutputsExpectedSize = 3;
-            assertEquals(svpFundTransactionUnsignedOutputsExpectedSize, svpFundTransactionUnsignedOutputs.size());
-            assertOneOutputIsChange(svpFundTransactionUnsignedOutputs);
-            assertOneOutputIsToProposedFederationWithExpectedAmount(svpFundTransactionUnsignedOutputs);
-            assertOneOutputIsToProposedFederationWithFlyoverPrefixWithExpectedAmount(svpFundTransactionUnsignedOutputs);
-        }
-
-        private void assertInputsAreFromActiveFederation() {
-            Script activeFederationScriptSig = PegTestUtils.createBaseInputScriptThatSpendsFromTheFederation(activeFederation);
-            List<TransactionInput> inputs = svpFundTransactionUnsigned.getInputs();
-
-            for (TransactionInput input : inputs) {
-                assertEquals(activeFederationScriptSig, input.getScriptSig());
-            }
-        }
-
-        private void assertOneOutputIsChange(List<TransactionOutput> transactionOutputs) {
-            Script activeFederationScriptPubKey = activeFederation.getP2SHScript();
-
-            Optional<TransactionOutput> changeOutputOpt = searchForOutput(transactionOutputs, activeFederationScriptPubKey);
-            assertTrue(changeOutputOpt.isPresent());
-        }
-
-        private void assertOneOutputIsToProposedFederationWithExpectedAmount(List<TransactionOutput> svpFundTransactionUnsignedOutputs) {
-            Script proposedFederationScriptPubKey = proposedFederation.getP2SHScript();
-
-            Optional<TransactionOutput> outputToProposedFederationOpt = searchForOutput(svpFundTransactionUnsignedOutputs, proposedFederationScriptPubKey);
-            assertTrue(outputToProposedFederationOpt.isPresent());
-
-            TransactionOutput outputToProposedFederation = outputToProposedFederationOpt.get();
-            assertEquals(outputToProposedFederation.getValue(), spendableValueFromProposedFederation);
-        }
-
-        private void assertOneOutputIsToProposedFederationWithFlyoverPrefixWithExpectedAmount(List<TransactionOutput> svpFundTransactionUnsignedOutputs) {
-            Script proposedFederationWithFlyoverPrefixScriptPubKey =
-                PegUtils.getFlyoverScriptPubKey(bridgeMainNetConstants.getProposedFederationFlyoverPrefix(), proposedFederation.getRedeemScript());
-
-            Optional<TransactionOutput> outputToProposedFederationWithFlyoverPrefixOpt = searchForOutput(
-                svpFundTransactionUnsignedOutputs,
-                proposedFederationWithFlyoverPrefixScriptPubKey
-            );
-            assertTrue(outputToProposedFederationWithFlyoverPrefixOpt.isPresent());
-
-            TransactionOutput outputToProposedFederationWithFlyoverPrefix = outputToProposedFederationWithFlyoverPrefixOpt.get();
-            assertEquals(outputToProposedFederationWithFlyoverPrefix.getValue(), spendableValueFromProposedFederation);
-        }
-
-        private Optional<TransactionOutput> searchForOutput(List<TransactionOutput> transactionOutputs, Script outputScriptPubKey) {
-            return transactionOutputs.stream()
-                .filter(output -> output.getScriptPubKey().equals(outputScriptPubKey))
-                .findFirst();
-        }
-
-        private void assertPegoutWasAddedToPegoutsWaitingForConfirmations(PegoutsWaitingForConfirmations pegoutsWaitingForConfirmations, Sha256Hash pegoutTransactionHash, Keccak256 releaseCreationTxHash) {
-            Set<PegoutsWaitingForConfirmations.Entry> pegoutEntries = pegoutsWaitingForConfirmations.getEntries();
-            Optional<PegoutsWaitingForConfirmations.Entry> pegoutEntry = pegoutEntries.stream()
-                .filter(entry -> entry.getBtcTransaction().getHash().equals(pegoutTransactionHash) &&
-                    entry.getPegoutCreationRskBlockNumber().equals(rskExecutionBlock.getNumber()) &&
-                    entry.getPegoutCreationRskTxHash().equals(releaseCreationTxHash))
-                .findFirst();
-            assertTrue(pegoutEntry.isPresent());
-        }
-
-        private void assertPegoutTxSigHashWasSaved(BtcTransaction pegoutTransaction) {
-            Optional<Sha256Hash> pegoutTxSigHashOpt = BitcoinUtils.getFirstInputSigHash(pegoutTransaction);
-            assertTrue(pegoutTxSigHashOpt.isPresent());
-
-            Sha256Hash pegoutTxSigHash = pegoutTxSigHashOpt.get();
-            assertTrue(bridgeStorageProvider.hasPegoutTxSigHash(pegoutTxSigHash));
-        }
-
-        private void assertLogReleaseRequested(Keccak256 releaseCreationTxHash, Sha256Hash pegoutTransactionHash, Coin requestedAmount) {
-            byte[] releaseCreationTxHashSerialized = releaseCreationTxHash.getBytes();
-            byte[] pegoutTransactionHashSerialized = pegoutTransactionHash.getBytes();
-            List<DataWord> encodedTopics = getEncodedTopics(releaseRequestedEvent, releaseCreationTxHashSerialized, pegoutTransactionHashSerialized);
-
-            byte[] encodedData = getEncodedData(releaseRequestedEvent, requestedAmount.getValue());
-
-            assertEventWasEmittedWithExpectedTopics(encodedTopics);
-            assertEventWasEmittedWithExpectedData(encodedData);
-        }
-
-        private void assertLogPegoutTransactionCreated(BtcTransaction pegoutTransaction) {
-            Sha256Hash pegoutTransactionHash = pegoutTransaction.getHash();
-            byte[] pegoutTransactionHashSerialized = pegoutTransactionHash.getBytes();
-            List<DataWord> encodedTopics = getEncodedTopics(pegoutTransactionCreatedEvent, pegoutTransactionHashSerialized);
-
-            List<Coin> outpointValues = extractOutpointValues(pegoutTransaction);
-            byte[] serializedOutpointValues = UtxoUtils.encodeOutpointValues(outpointValues);
-            byte[] encodedData = getEncodedData(pegoutTransactionCreatedEvent, serializedOutpointValues);
-
-            assertEventWasEmittedWithExpectedTopics(encodedTopics);
-            assertEventWasEmittedWithExpectedData(encodedData);
-        }
-
-        private List<DataWord> getEncodedTopics(CallTransaction.Function bridgeEvent, Object... args) {
-            byte[][] encodedTopicsInBytes = bridgeEvent.encodeEventTopics(args);
-            return LogInfo.byteArrayToList(encodedTopicsInBytes);
-        }
-
-        private byte[] getEncodedData(CallTransaction.Function bridgeEvent, Object... args) {
-            return bridgeEvent.encodeEventData(args);
-        }
-
-        private void assertEventWasEmittedWithExpectedTopics(List<DataWord> expectedTopics) {
-            Optional<LogInfo> topicOpt = logs.stream()
-                .filter(log -> log.getTopics().equals(expectedTopics))
-                .findFirst();
-            assertTrue(topicOpt.isPresent());
-        }
-
-        private void assertEventWasEmittedWithExpectedData(byte[] expectedData) {
-            Optional<LogInfo> data = logs.stream()
-                .filter(log -> Arrays.equals(log.getData(), expectedData))
-                .findFirst();
-            assertTrue(data.isPresent());
-        }
-    }
-
-    @Nested
     @Tag("LockingCap")
     class LockingCapTest {
